Bugfix & Code reformat
git-svn-id: http://escidev5.fiz-karlsruhe.de:8888/prototype/eSciDocCore@42980 5c398c58-3d0a-0410-a38c-d0825f6816e0

+++ utilities/src/main/java/org/esidoc/core/utils/io/FileUtils.java
@@ -12,7 +12,7 @@ import java.util.Random;
 public final class FileUtils {
 
     private static final Random RANDOM = new Random();
-    private static final int RETRY_SLEEP_MILLIS = 10;
+    private static final long RETRY_SLEEP_MILLIS = 10;
 
     private static File defaultTempDirectory;
